back-end/requirements.txt
# API dependencies
fastapi
uvicorn

# LLM dependencies
PyPDF2==3.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
langchain==0.0.154
# Peer deps of langchain
openai==0.28.1
tiktoken